stub in signup view and toggle when button tapped
